provide default decoder as public api
