+++ lib/index.js
@@ -2,3 +2,4 @@ export { encodeVLQ, decodeVLQ } from './vlq';
 export { default as IntBufferReader } from './int-buffer-reader';
 export { default as IntBufferWriter } from './int-buffer-writer';
 export { default as MappingsDecoder } from './mappings-decoder';
+export { default as Decoder } from './decoder';
